[OPIK-6245] [FE] feat: show Ollie version in AppDebugInfo
Surface the assistant pod's manifest version next to OPIK_VERSION
when Comet Debugger Mode (Meta/Ctrl+Shift+.) is on. Reuses the
existing react-query cache (no extra network request) and renders
nothing when the assistant is disabled or the pod isn't ready.

- Extract manifest fetch into shared useAssistantManifest hook
  (consumed by both AssistantSidebar and the new AssistantDebugInfo)
- New AssistantDebugInfo plugin component (V2 only, comet/dev folders)
- Register AssistantDebugInfo slot in PluginsStore so OSS prod renders
  nothing
- Polish AppDebugInfo / AppNetworkStatus sizing (size-4 icons, size-3
  copy, comet-body-xs-accented, size-1.5 status dots, gap-3 between
  blocks)
