+++ apps/opik-frontend/src/plugins/development/AssistantDebugInfo.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/plugins/comet/AssistantDebugInfo";
