+++ apps/opik-frontend/src/v2/layout/AppDebugInfo/AppDebugInfo.tsx
@@ -7,12 +7,16 @@ import OpikIcon from "@/icons/opik.svg?react";
 import copy from "clipboard-copy";
 import { Copy } from "lucide-react";
 import { modifierKey } from "@/lib/utils";
+import usePluginsStore from "@/store/PluginsStore";
 
 const DEBUGGER_MODE_KEY = "comet-debugger-mode"; // Same key used in EM for consistency
 
 const AppDebugInfo = () => {
   const [showAppDebugInfo, setShowAppDebugInfo] = useState(
     () => localStorage.getItem(DEBUGGER_MODE_KEY) === "true",
+  );
+  const AssistantDebugInfo = usePluginsStore(
+    (state) => state.AssistantDebugInfo,
   );
 
   // Keyboard shortcut handler for debugger mode: Meta/Ctrl + Shift + . (all pressed simultaneously)
@@ -30,27 +34,27 @@ const AppDebugInfo = () => {
 
   return (
     showAppDebugInfo && (
-      <>
-        <div className="flex items-center">
-          <AppNetworkStatus />
-        </div>
+      <div className="flex items-center gap-3">
+        <AppNetworkStatus />
 
         {APP_VERSION && (
           <div
-            className="flex items-center gap-2"
+            className="flex items-center gap-1"
             onClick={() => {
               copy(APP_VERSION);
-              toast({ description: "Successfully copied version" });
+              toast({ description: "Successfully copied Opik version" });
             }}
           >
-            <span className="comet-body-s-accented flex items-center gap-2 truncate">
-              <OpikIcon className="size-5" />
+            <span className="comet-body-xs-accented flex items-center gap-1 truncate">
+              <OpikIcon className="size-4" />
               OPIK VERSION {APP_VERSION}
             </span>
-            <Copy className="size-4 shrink-0" />
+            <Copy className="size-3 shrink-0" />
           </div>
         )}
-      </>
+
+        {AssistantDebugInfo && <AssistantDebugInfo />}
+      </div>
     )
   );
 };
